requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-dotenv==1.0.0
openai==1.54.0
pydantic==2.9.0
langgraph==0.2.45
langchain-openai==0.2.8
langchain-core==0.3.15